Added OpenMP; Lossless intra encoding fully functional

+++ src/io/Golomb.cpp
@@ -9,14 +9,16 @@ using namespace std;
 Golomb::Golomb(const std::string &filePath, std::ios_base::openmode mode) {
     bs = new BitStream(filePath, mode);
     filepath = filePath;
-    localStream=true;
+    localStream = true;
 }
 
 Golomb::Golomb(BitStream *bitstream) {
     bs = bitstream;
 }
 
-Golomb::~Golomb() { if(localStream){delete bs;}}
+Golomb::~Golomb() {
+    if (localStream) { delete bs; }
+}
 
 void Golomb::_set_m(int m_) { m = m_; }
 int Golomb::_get_m() const { return m; }
@@ -24,24 +26,24 @@ int Golomb::_get_m() const { return m; }
 
 int Golomb::decode() {
     if (m <= 0) {
-        m = bs->readBits(8*sizeof(int));
+        m = bs->readBits(8 * sizeof(int));
     }
-    int sign=(bs->readBit()==0)?1:-1;
+    int sign = (bs->readBit() == 0) ? 1 : -1;
     int q = readUnary();
     int r = readBinaryTrunc();
-    return sign*(q * m + r);
+    return sign * (q * m + r);
 }
 
 void Golomb::encode(int n) {
     if (m <= 0) {
         throw std::invalid_argument("value of m unknown or invalid");
     }
-    if(n<0){
+    if (n < 0) {
         bs->writeBit(1);
-    }else{
+    } else {
         bs->writeBit(0);
     }
-    n=abs(n);
+    n = abs(n);
     int r = n % m;
     int q = n / m;
     writeUnary(q);
@@ -51,7 +53,7 @@ void Golomb::encode(int n) {
 void Golomb::encode(int n, int m_) {
     if (m <= 0) {
         m = m_;
-        bs->writeBits(m_, 8*sizeof(int));
+        bs->writeBits(m_, 8 * sizeof(int));
     }
     encode(n);
 }
@@ -61,7 +63,7 @@ int Golomb::readUnary() {
     int q = 0;
     int bit_ = -1;
     while (bit_ != 0) {
-        bit_=bs->readBit();
+        bit_ = bs->readBit();
         q++;
     }
     q--;//the last one (0) doesn't count
@@ -77,7 +79,7 @@ void Golomb::writeUnary(int n) {
 
 int Golomb::readBinaryTrunc() {
     int k = floor(log2(m));
-    int u = pow(k + 1, 2) - m;
+    int u = (int) pow(k + 1, 2) - m;
     int k_bits = bs->readBits(k);
     if (k_bits < u) {
         return k_bits;
@@ -88,7 +90,7 @@ int Golomb::readBinaryTrunc() {
 
 void Golomb::writeBinaryTrunc(int n) {
     int k = floor(log2(m));
-    int u = pow(k + 1, 2) - m;
+    int u = (int) pow(k + 1, 2) - m;
     if (n < u)
         bs->writeBits(n, k);
     else
